refactor(sync): split storage snapshots and configuration
